+++ src/JsonMapper.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace JsonMapper;
 
-use JsonException;
 use JsonMapper\Dto\NamedMiddleware;
 use JsonMapper\Exception\TypeError;
 use JsonMapper\ValueObjects\PropertyMap;
@@ -90,6 +89,21 @@ class JsonMapper implements JsonMapperInterface
         return $this;
     }
 
+    public function mapToClass(\stdClass $json, string $class)
+    {
+        if (! \class_exists($class)) {
+            throw TypeError::forArgument(__METHOD__, 'class-string', $class, 2, '$class');
+        }
+
+        $propertyMap = new PropertyMap();
+
+        $handler = $this->resolve();
+        $wrapper = new ObjectWrapper(null, $class);
+        $handler($json, $wrapper, $propertyMap, $this);
+
+        return $wrapper->getObject();
+    }
+
     public function mapObject(\stdClass $json, $object)
     {
         if (! \is_object($object)) {
@@ -117,6 +131,34 @@ class JsonMapper implements JsonMapperInterface
         }
 
         return $results;
+    }
+
+    public function mapToClassArray(array $json, string $class): array
+    {
+        if (! \class_exists($class)) {
+            throw TypeError::forArgument(__METHOD__, 'class-string', $class, 2, '$class');
+        }
+
+        return array_map(
+            function (\stdClass $value) use ($class) {
+                return $this->mapToClass($value, $class);
+            },
+            $json
+        );
+    }
+
+    public function mapToClassFromString(string $json, string $class)
+    {
+        if (! \class_exists($class)) {
+            throw TypeError::forArgument(__METHOD__, 'class-string', $class, 2, '$class');
+        }
+
+        $data = $this->decodeJsonString($json);
+        if (! $data instanceof \stdClass) {
+            throw new \RuntimeException('Provided string is not a json encoded object');
+        }
+
+        return $this->mapToClass($data, $class);
     }
 
     public function mapObjectFromString(string $json, $object)
@@ -155,6 +197,20 @@ class JsonMapper implements JsonMapperInterface
         }
 
         return $results;
+    }
+
+    public function mapToClassArrayFromString(string $json, string $class): array
+    {
+        if (! \class_exists($class)) {
+            throw TypeError::forArgument(__METHOD__, 'class-string', $class, 2, '$class');
+        }
+
+        $data = $this->decodeJsonString($json);
+        if (! \is_array($data)) {
+            throw new \RuntimeException('Provided string is not a json encoded array');
+        }
+
+        return $this->mapToClassArray($data, $class);
     }
 
     /** @return \stdClass|\stdClass[] */