+++ src/JsonMapperBuilder.php
@@ -8,9 +8,11 @@ use JsonMapper\Cache\ArrayCache;
 use JsonMapper\Dto\NamedMiddleware;
 use JsonMapper\Enums\TextNotation;
 use JsonMapper\Exception\BuilderException;
+use JsonMapper\Handler\FactoryRegistry;
 use JsonMapper\Handler\PropertyMapper;
 use JsonMapper\Middleware\Attributes\Attributes;
 use JsonMapper\Middleware\CaseConversion;
+use JsonMapper\Middleware\Constructor\Constructor;
 use JsonMapper\Middleware\Debugger;
 use JsonMapper\Middleware\DocBlockAnnotations;
 use JsonMapper\Middleware\FinalCallback;
@@ -145,6 +147,15 @@ class JsonMapperBuilder
         bool $onlyApplyCallBackOnTopLevel = true
     ): JsonMapperBuilder {
         return $this->withMiddleware(new FinalCallback($callback, $onlyApplyCallBackOnTopLevel), FinalCallback::class);
+    }
+
+    public function withObjectConstructorMiddleware(FactoryRegistry $factoryRegistry): JsonMapperBuilder
+    {
+        // @TODO Add the registry from the builder context?
+        return $this->withMiddleware(
+            new Constructor($factoryRegistry), // How to get logical fallback the factory registry
+            Constructor::class
+        );
     }
 
     public function withMiddleware(callable $middleware, ?string $name = null): JsonMapperBuilder