+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Constructor middleware; Add support for mapping to class name  [PR#141](https://github.com/JsonMapper/JsonMapper/pull/141)
 
 ## [2.13.0] - 2022-08-17
 ### Added